Update traffic analysis
Use link capacity as the filter (instead of actual traffic volume), such that direct comparison between 2 different setup can be made.
