+++ contribs/application/src/main/java/org/matsim/application/analysis/travelTimeValidation/TravelTimeAnalysis.java
@@ -1,4 +1,4 @@
-package org.matsim.application.analysis;
+package org.matsim.application.analysis.travelTimeValidation;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -29,7 +29,6 @@ import java.time.LocalDate;
 import java.time.temporal.TemporalAdjusters;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Random;
 import java.util.Set;
 import java.util.function.Predicate;
